refactor: split SidePanel per-tab — slim router + 5 tab modules (#306)
SidePanel.tsx drops from 1129 LOC to 139 (slim router that imports
the Tabs strip + dispatches by current tab). Each tab moves to its
own module under components/sidepanel/, with cross-tab helpers
(basename, layerAccent, formatScore, StatRow, SectionHeading,
ScoreButton) hoisted into sidepanel/shared.tsx.

Public surface unchanged: SidePanel, SidePanelProps, and Issue still
export from ./SidePanel.

  src/components/sidepanel/
    shared.tsx       cross-tab helpers
    InfoTab.tsx      ScoreBlock + WelcomeCard + Stats + HistoryScrubber
                     + AutoLayersBanner + formatTimestamp
    FilesTab.tsx     FilesTab + TreeRow
    LevelsTab.tsx
    IssuesTab.tsx    + Issue type + IssueKind + buildIssues + labelFor
                     + kindClass + cycleSubtitle + humaniseFlag
    RulesTab.tsx
  src/components/SidePanel.tsx   slim router (Tabs + dispatch)

Verified: pnpm build clean, 30/30 Playwright e2e green, 308/308
cargo tests green, cargo clippy --workspace -- -D warnings clean.
